+++ storage-core/src/chunkserver/external/mod.rs
@@ -1,0 +1,5 @@
+mod chunkserver_definition;
+mod chunkserver_impl;
+mod server_impl;
+
+pub use self::chunkserver_definition::ChunkserverExternal;
