+++ storage-core/src/chunkserver/internal/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod chunkserver_definition;
+pub(crate) mod chunkserver_impl;
+pub(crate) mod server_impl;
+
+pub use self::chunkserver_definition::ChunkserverInternal;
