+++ storage-core/src/chunkserver/mod.rs
@@ -1,3 +1,7 @@
-pub(crate) mod external_chunkserver;
 pub(crate) mod config;
 pub(crate) mod setup;
+pub(crate) mod external;
+pub(crate) mod internal;
+
+use self::external::ChunkserverExternal;
+use self::internal::ChunkserverInternal;
