+++ storage-core/src/client/main.rs
@@ -1,3 +1,5 @@
+mod client;
+
 use std::{
     fs,
     io::{self, Write},
@@ -109,8 +111,8 @@ async fn run(options: Opt) -> Result<()> {
         .await
         .map_err(|e| anyhow!("failed to open stream: {}", e))?;
     if rebind {
-        let socket = std::net::UdpSocket::bind("[::]:0").unwrap();
-        let addr = socket.local_addr().unwrap();
+        let socket = std::net::UdpSocket::bind("[::]:0")?;
+        let addr = socket.local_addr()?;
         eprintln!("rebinding to {addr}");
         endpoint.rebind(socket).expect("rebind failed");
     }
@@ -118,7 +120,7 @@ async fn run(options: Opt) -> Result<()> {
     send.write_all(request.as_bytes())
         .await
         .map_err(|e| anyhow!("failed to send request: {}", e))?;
-    send.finish().unwrap();
+    send.finish()?;
     let response_start = Instant::now();
     eprintln!("request sent at {:?}", response_start - start);
     let resp = recv
@@ -131,8 +133,8 @@ async fn run(options: Opt) -> Result<()> {
         duration,
         resp.len() as f32 / (duration_secs(&duration) * 1024.0)
     );
-    io::stdout().write_all(&resp).unwrap();
-    io::stdout().flush().unwrap();
+    io::stdout().write_all(&resp)?;
+    io::stdout().flush()?;
     conn.close(0u32.into(), b"done");
 
     // Give the server a fair chance to receive the close packet
