+++ storage-core/src/common/mod.rs
@@ -1,15 +1,16 @@
 //#![cfg(any(feature = "rustls-aws-lc-rs", feature = "rustls-ring"))]
 //! Commonly used code in most examples.
 
+mod chunk_send;
 pub mod message_payload;
 mod messages;
 mod server;
 mod types;
 
+pub use chunk_send::ChunkserverLocation;
 pub use message_payload::*;
 pub use messages::*;
 pub use server::{CertificateProvider, QuicServer, certificate_provider};
-pub use types::ChunkserverLocation;
 
 use quinn::{ClientConfig, Endpoint, ServerConfig};
 use rustls::pki_types::{CertificateDer, PrivatePkcs8KeyDer};
